# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Tech Events Hub API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiError {
  message: string;
}

export interface ScrapeRequest {
  /** Optional source slug. Omit to run all scrapers. */
  source?: string;
}

export type EventType = (typeof EventType)[keyof typeof EventType];

export const EventType = {
  hackathon: "hackathon",
  webinar: "webinar",
  workshop: "workshop",
  other: "other",
} as const;

export type EventMode = (typeof EventMode)[keyof typeof EventMode];

export const EventMode = {
  online: "online",
  offline: "offline",
  hybrid: "hybrid",
  unknown: "unknown",
} as const;

export interface Event {
  id: number;
  title: string;
  platform: string;
  type: EventType;
  url: string;
  image?: string | null;
  startDate?: string | null;
  endDate?: string | null;
  mode: EventMode;
  tags: string[];
  organizer?: string | null;
  location?: string | null;
  prize?: string | null;
  description?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface EventListResponse {
  events: Event[];
  total: number;
}

export interface PlatformCount {
  platform: string;
  count: number;
}

export interface TypeCount {
  type: string;
  count: number;
}

export interface ModeCount {
  mode: string;
  count: number;
}

export interface TagCount {
  tag: string;
  count: number;
}

export interface StatsOverview {
  totalEvents: number;
  hackathons: number;
  webinars: number;
  workshops: number;
  upcomingThisWeek: number;
  byPlatform: PlatformCount[];
  byType: TypeCount[];
  byMode: ModeCount[];
  lastScrapedAt: string | null;
}

export type ScrapeJobStatus =
  (typeof ScrapeJobStatus)[keyof typeof ScrapeJobStatus];

export const ScrapeJobStatus = {
  running: "running",
  success: "success",
  error: "error",
} as const;

export interface ScrapeJob {
  id: number;
  source: string;
  status: ScrapeJobStatus;
  eventsFound: number;
  eventsUpserted: number;
  errorMessage?: string | null;
  startedAt: string;
  finishedAt?: string | null;
}

export interface ScrapeResult {
  jobs: ScrapeJob[];
}

export type ListEventsParams = {
  type?: ListEventsType;
  platform?: string;
  mode?: ListEventsMode;
  tag?: string;
  search?: string;
  /**
   * @minimum 1
   * @maximum 200
   */
  limit?: number;
  /**
   * @minimum 0
   */
  offset?: number;
};

export type ListEventsType =
  (typeof ListEventsType)[keyof typeof ListEventsType];

export const ListEventsType = {
  hackathon: "hackathon",
  webinar: "webinar",
  workshop: "workshop",
  all: "all",
} as const;

export type ListEventsMode =
  (typeof ListEventsMode)[keyof typeof ListEventsMode];

export const ListEventsMode = {
  online: "online",
  offline: "offline",
  hybrid: "hybrid",
  all: "all",
} as const;

export type GetTrendingTagsParams = {
  /**
   * @minimum 1
   * @maximum 50
   */
  limit?: number;
};

export type GetTrendingTags200 = {
  tags: TagCount[];
};

export type GetUpcomingDeadlinesParams = {
  /**
   * @minimum 1
   * @maximum 50
   */
  limit?: number;
};

export type ListScrapeJobsParams = {
  /**
   * @minimum 1
   * @maximum 50
   */
  limit?: number;
};

export type ListScrapeJobs200 = {
  jobs: ScrapeJob[];
};