# lib/api-zod/src/generated/types/platformCount.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Tech Events Hub API
 * OpenAPI spec version: 0.1.0
 */

export interface PlatformCount {
  platform: string;
  count: number;
}