# lib/api-zod/src/generated/types/scrapeJob.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Tech Events Hub API
 * OpenAPI spec version: 0.1.0
 */
import type { ScrapeJobStatus } from "./scrapeJobStatus";

export interface ScrapeJob {
  id: number;
  source: string;
  status: ScrapeJobStatus;
  eventsFound: number;
  eventsUpserted: number;
  errorMessage?: string | null;
  startedAt: string;
  finishedAt?: string | null;
}