# lib/api-zod/src/generated/types/statsOverview.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Tech Events Hub API
 * OpenAPI spec version: 0.1.0
 */
import type { ModeCount } from "./modeCount";
import type { PlatformCount } from "./platformCount";
import type { TypeCount } from "./typeCount";

export interface StatsOverview {
  totalEvents: number;
  hackathons: number;
  webinars: number;
  workshops: number;
  upcomingThisWeek: number;
  byPlatform: PlatformCount[];
  byType: TypeCount[];
  byMode: ModeCount[];
  lastScrapedAt: string | null;
}